pyproject.toml
[project]
name = "elt-job"
version = "0.1.0"
description = ""
authors = [
    {name = "RenataAraujo",email = "renata.aesilva@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.5"
dependencies = [
    "pandas (>=3.0.0,<4.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "snowflake-sqlalchemy (>=1.8.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "dbt-core (>=1.11.3,<2.0.0)",
    "dbt-snowflake (>=1.11.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
